Change ASCRead behaviour By default data are tarnsformed into points not polygons (option)
In the future :
We must add new options 
- to set the type of z value (currently double for points..., int for polygons)
- to specify if nodata point must be imported or not

+++ h2gis-functions/src/main/java/org/h2gis/functions/io/asc/AscReaderDriver.java
@@ -180,6 +180,7 @@ public class AscReaderDriver {
      * @param inputStream
      * @param progress
      * @param tableReference
+     * @param srid the espg code of the input file
      * @throws SQLException
      * @throws IOException
      */
@@ -247,7 +248,7 @@ public class AscReaderDriver {
                             if (data != noData) {
                                 preparedStatement.setObject(2, data);
                             } else {
-                                preparedStatement.setNull(2, Types.INTEGER);
+                                preparedStatement.setNull(2, Types.DOUBLE);
                             }
                             preparedStatement.addBatch();
                             batchSize++;
